+++ src/test/test_solver/test_base_serial_solver.cpp
@@ -7,16 +7,16 @@
 bool test_base_serial_solver()
 {
     bool all_passed = true;
-    auto initial_cond_function = [](double x, double y)
-    { return std::complex<double>{1 * std::exp(-(x * x + y * y) / (9)), 0}; };
+    auto initial_cond_function = [](float x, float y)
+    { return std::complex<float>{1 * std::exp(-(x * x + y * y) / (9)), 0}; };
 
     auto initial_condition = new InitialCondition(initial_cond_function);
 
-    std::function<double(double, double)> potential = [](double x, double y)
+    std::function<float(float, float)> potential = [](float x, float y)
     {
-        return (double)0.5 * (x * x + y * y);
+        return (float)0.5 * (x * x + y * y);
     };
-    double g = 1;
+    float g = 1;
     BaseDomain *domain = new BaseDomain(11, 11, 0, 10, 11);
     try{
         BaseSolver solver = BaseSolver(potential, g);
